components folder restructuring
