+++ wsc-vite-app/src/components/event/Event.jsx
