+++ wsc-vite-app/src/components/footer/Footer.jsx
